+++ jagobelanja-api/src/main/java/com/jagobelanja/service/CategoryService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface CategoryService {
     List<Category> getAllCategories();
-    Category getCategoryById(long id);
+    Category getCategoryById(int id);
     Category createCategory(Category category);
     Category updateCategory(Category category);
-    void deleteCategory(long id);
+    void deleteCategory(int id);
 }
